requirements.txt
beautifulsoup4
requests
flask
pypdf
pdfplumber
PyJWT
